week-0/databases/Simulate-PineCone/requirements.txt
streamlit==1.39.0
pinecone-client==3.0.3
plotly==5.24.1
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4
seaborn==0.13.2
matplotlib==3.9.2
python-dotenv==1.0.1
requests==2.32.3